requirements.txt
pyserial==3.5
customtkinter==5.2.1
schedule==1.2.1